Imitation Training Properly Setup
Now we can run the recorder to record our actions and outputs.
After that we can train the agents with the recording.
With the inference we can see what the agents have learned with the recordings.
We could then run the Reinforcement learning on the same data set to improve the behavior's even further